+++ packages/amplify-category-analytics/src/utils/kinesis-helper.ts
@@ -1,5 +1,5 @@
 import { $TSAny, $TSContext, $TSMeta, AmplifyCategories, AmplifySupportedService, open } from 'amplify-cli-core';
-import { printer } from 'amplify-prompts';
+import { printer } from '@aws-amplify/amplify-prompts';
 
 /**
  * opens resource in AWS console