+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/auth-stack-builder/auth-stack-transform.ts
@@ -14,8 +14,8 @@ import {
   stateManager,
   Template,
   writeCFNTemplate,
-} from 'amplify-cli-core';
-import { formatter } from 'amplify-prompts';
+} from '@aws-amplify/amplify-cli-core';
+import { formatter } from '@aws-amplify/amplify-prompts';
 import * as cdk from 'aws-cdk-lib';
 import * as fs from 'fs-extra';
 import _ from 'lodash';