+++ packages/amplify-category-function/src/provider-utils/awscloudformation/service-walkthroughs/secretValuesWalkthrough.ts
@@ -3,7 +3,7 @@
  */
 
 import { ResourceName } from 'amplify-cli-core';
-import { FunctionParameters, removeSecret, SecretDeltas, setSecret } from 'amplify-function-plugin-interface';
+import { FunctionParameters, removeSecret, SecretDeltas, setSecret } from '@aws-amplify/amplify-function-plugin-interface';
 import inquirer from 'inquirer';
 import { getExistingSecrets, hasExistingSecrets } from '../secrets/secretDeltaUtilities';
 import { getStoredEnvironmentVariables } from '../utils/environmentVariablesHelper';