+++ packages/amplify-category-geo/src/service-utils/resourceUtils.ts
@@ -1,4 +1,4 @@
-import { JSONUtilities, pathManager, $TSObject, stateManager, $TSContext } from 'amplify-cli-core';
+import { JSONUtilities, pathManager, $TSObject, stateManager, $TSContext } from '@aws-amplify/amplify-cli-core';
 import { category, authCategoryName } from '../constants';
 import path from 'path';
 import _ from 'lodash';
@@ -9,7 +9,7 @@ import os from 'os';
 import { getMapIamPolicies } from './mapUtils';
 import { getPlaceIndexIamPolicies } from './placeIndexUtils';
 import { getGeofenceCollectionIamPolicies } from './geofenceCollectionUtils';
-import { printer } from 'amplify-prompts';
+import { printer } from '@aws-amplify/amplify-prompts';
 
 // Merges other with existing in a non-destructive way.
 // Specifically, scalar values will not be overwritten