+++ packages/amplify-category-geo/src/service-walkthroughs/placeIndexWalkthrough.ts
@@ -2,12 +2,12 @@ import { v4 as uuid } from 'uuid';
 import { merge } from '../service-utils/resourceUtils';
 import { DataSourceIntendedUse, PlaceIndexParameters } from '../service-utils/placeIndexParams';
 import { apiDocs, ServiceName } from '../service-utils/constants';
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext } from '@aws-amplify/amplify-cli-core';
 import { getCurrentPlaceIndexParameters } from '../service-utils/placeIndexUtils';
 import { getGeoServiceMeta, updateDefaultResource, checkGeoResourceExists, getGeoResources } from '../service-utils/resourceUtils';
 import { resourceAccessWalkthrough, dataProviderWalkthrough, getServiceFriendlyName, defaultResourceQuestion } from './resourceWalkthrough';
 import { DataProvider } from '../service-utils/resourceParams';
-import { printer, formatter, prompter, alphanumeric } from 'amplify-prompts';
+import { printer, formatter, prompter, alphanumeric } from '@aws-amplify/amplify-prompts';
 
 const searchServiceFriendlyName = getServiceFriendlyName(ServiceName.PlaceIndex);
 /**