+++ packages/amplify-category-notifications/src/notifications-api.ts
@@ -7,7 +7,7 @@ import {
   INotificationsResourceMeta,
   stateManager,
 } from 'amplify-cli-core';
-import { printer } from 'amplify-prompts';
+import { printer } from '@aws-amplify/amplify-prompts';
 import { IChannelAPIResponse, ChannelAction, INotificationsConfigStatus } from './channel-types';
 import { PinpointName } from './pinpoint-name';
 import { disableNotificationsChannel, enableNotificationsChannel, getChannelAvailability } from './notifications-backend-cfg-channel-api';