+++ packages/amplify-provider-awscloudformation/src/resource-package/resource-packager.ts
@@ -322,7 +322,7 @@ export abstract class ResourcePackager {
 
   // eslint-disable-next-line class-methods-use-this
   private getCfnTemplatePathsForResource(resource: ResourceDefinition): string[] {
-    const { cfnFiles } = getCfnFiles(resource.category, resource.resourceName, {
+    const { cfnFiles } = getCfnFiles(resource.category, resource.resourceName, false, {
       absolute: true,
     });
     return cfnFiles;